+++ docs/change_log.md
@@ -1,6 +1,27 @@
 # *bixverse* change log
 
-*Last update: 06.08.2025* </br>
+*Last update: 04.09.2025* </br>
+
+### Version **0.0.2.1**
+
+#### New features
+
+- Addition of constrained PageRank implementation.
+- Improved synthetic data for correlation-based gene module detections.
+- Improved tests for ICA, CoReMo and graph-based correlation modules.
+- Various distance calculations accelerated in Rust.
+- The Rust-based implementations of personalised PageRank have now the option
+to use weighted graphs.
+
+#### Bug fixes, documentation updates
+
+- Fixed bugs for the graph-based correlation methods.
+- Fixed export bugs for some of the plotting functions.
+
+#### Breaking changes
+
+- Interface to the Rust-based personalised page-rank calculations can break due
+to the addition of a `weights` parameter.
 
 ### Version **0.0.2.0**
 
@@ -19,11 +40,11 @@ Additionally, permutation-based testing for diffusions is implemented.
 - Added reciprocal best hit method based on correlations. You will need to
 provide now an additional parameter to the class specifying if you want to use
 set similarity or correlation-based similarity.
-- Vignettes for various methods written and added. (More to come!)
+- Vignettes for various methods written and added.
 - GSVA and ssGSEA implemented in Rust.
 - Rust code reworked in different places to make some of the functions faster.
-- Mutual information and pointwise mutual information calculations in Rust.
-- Improvements to the synthetic data with additonal plotting functions.
+- Mutual information calculations in Rust.
+- Improvements to the synthetic data.
 
 #### Bug fixes, documentation updates